Yet another version of 2021 day 11
This uses integers (powers of 2) to represent the chemical elements
(rather than their names). So the members of the frozensets are integers
rather than strings.

This also uses a more conservative floor-state representation which only
considers any microchip-generator pairs to be equivalent. Each floor is
represented by a single integer: the number of microchip-generator pairs
plus the sum of the remaining microchips left-shifted by 4 plus the sum
of the remaining generators left-shifted by 12.